Fix translation download from crowdin
Unfortunately I am unable to get the 'gradlecrowdin' plugin to copy the downloaded files to the final destination. This commit has the plugin copy the contents to a temporary directory and than uses a copy task to copy matching files to the final destination.

In the long run the 'gradlecrowdin' plugin could be replaced by a custom implementation to skip the overhead.

Closes #132.