Remove unneeded position correction for legacy positions
Problematic positions are already corrected via the default position validation. Only if it is disabled, the legacy correction would be needed at all. However, if safety checks are disabled, I think that there should not be any changes to the original position. Also, relying on one validation algorithm only simplifies the handling and reduces the code platform implementations have to provide.

Addresses #146.