[Bukkit] Add a way to specify JDBC configuration options
For now this is limited to the configuration of MyWarp's main SQL DB in the 'config.yml'. There is no way (yet) to specify additional connection properties when using the import command. However, I think in these cases this is hardly needed as these are on time operations where a warning , e.g. about unsigned SSL certificates hardly hurts.

Closes #178.