Accept environment variables and properties for build options
Options, such as the build number, CI name or crowdin's API key can either be given as a property (using 'gradlew -Pkey=value') or as an environment variable. If both are present, the property takes priority. If non is present, the build script uses a default value.

This should also fix wrong build numbers when Travis attempts to publish artifacts on mywarp.github.com.