[Bukkit] Remove cancellation of interaction events for warp signs
The removed cancellation was added by 1b75d82 in order to fix 'Player moved wrongly!' messages. When testing, the removal of the cancellation does not result in such messages, perhaps because players are now teleported with TeleportCause.PLUGIN.

However, if the problem occurs again, cancellation might have to be added again.

Fixes #224.