Update Gradle to 7.0
The 'com.mendhak.gradlecrowdin' plugin is not compatible with Gradle 7 and had to be removed. Instead, we now download all translations by using Crowdin's API directly.

Unfortunately, the Crowdin's API v1 is scheduled to be removed at the end of 2021, so we may need to update the integration in the future.